Replace React Scripts testing with Vitest
