+++ package.json
@@ -14,6 +14,8 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/uuid": "^9.0.1",
     "jquery": "^3.6.3",
+    "jsdom": "^22.1.0",
+    "playwright": "^1.40.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -26,8 +28,7 @@
   "scripts": {
     "start": "vite",
     "build": "vite build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "test": "vitest",
     "dev": "vite"
   },
   "eslintConfig": {
@@ -52,11 +53,13 @@
   "devDependencies": {
     "@types/jquery": "^3.5.16",
     "@vitejs/plugin-react": "^3.1.0",
+    "@vitest/browser": "^0.34.6",
     "prettier": "2.8.8",
     "prettier-plugin-java": "^2.1.0",
     "sass": "^1.58.3",
     "vite": "^4.2.1",
     "vite-plugin-svgr": "^2.4.0",
-    "vite-tsconfig-paths": "^3.5.0"
+    "vite-tsconfig-paths": "^3.5.0",
+    "vitest": "^0.34.6"
   }
 }
