+++ src/test/bindings/ir.test.ts
@@ -1,11 +1,13 @@
 import { sequence } from '../../bindings/ir'
+import { test, expect } from 'vitest'
 
 test("basic", () => {
   const s = sequence((s, p1, p2) => {
     s.run("Drive Base", "Command X", p1, p2, "hardcoded")
   });
 
-  expect(s.params).toEqual([{ name: 'p1', original: null }, { name: 'p2', original: null }]);
+  const b = [{ name: 'p1', original: null }, {name: 'p2', original: null }]
+  expect(s.params.every((obj, i) => obj.name === b[i].name && obj.original === b[i].original)).to.be.true
 });
 
 test("nesting", () => {
