+++ vitest.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+    test: {
+        environment: 'jsdom',
+        browser: {
+            enabled: false,
+            provider: 'playwright',
+            name: 'chromium',
+        },
+    },
+})
